Changes aprioriRisk to int.
Instead of worrying about validating the text box, just rely on the
property being an int. If anything besides an int is entered, the box
border turns red.

+++ DataBinding-Attempt/RiskAssessmentModel.cs
@@ -90,8 +90,8 @@
         };
 
         // a priori Risk
-        private double _aprioriRisk;
-        public double aprioriRisk
+        private int _aprioriRisk;
+        public int aprioriRisk
         {
             get { return _aprioriRisk; }
             set
